小更新，新增了点赞、阅读量逻辑、以及热门文章。删除了无用的示例内容，修复了一些作者显示只显示id而非用户名字的地方。

+++ src/main/java/com/murasame/domain/dto/CommentDTO.java
@@ -9,7 +9,7 @@ public class CommentDTO {
 	Long id;
 	Long b_id;
 	Long parent_cid;
-	Integer u_id;
+	Long u_id;
 	String content;
 	LocalDateTime created_at;
 }
